Refactoring into two separate components
